Move EventHubConnectionStringProperties into models folder.
